CORE-15575: initial implementation of forex sample
