+++ kotlin-samples/Advanced/foreign-exchange/.ci/nightly/JenkinsfileSnykScan
