+++ kotlin-samples/Advanced/foreign-exchange/config/log4j2.xml
