+++ kotlin-samples/Advanced/foreign-exchange/gradle.properties
