--- kotlin-samples/Basic/primenumber/contracts/src/main/kotlin/com/r3/developers/samples/primenumber/PrimeContract.kt
@@ -1,43 +1,0 @@
-package com.r3.developers.samples.primenumber
-
-import net.corda.v5.base.exceptions.CordaRuntimeException
-import net.corda.v5.ledger.utxo.Contract
-import net.corda.v5.ledger.utxo.transaction.UtxoLedgerTransaction
-import java.lang.IllegalArgumentException
-
-class PrimeContract: Contract {
-
-    internal companion object {
-        const val REQUIRE_SINGLE_COMMAND = "Requires a single PrimeCommands command."
-        const val UNRECOGNISED_COMMAND = "Incorrect type of PrimeCommands."
-        const val NON_POSITIVE = "The prime number generated is not greater than 0"
-        const val NON_MATCHING = "The prime in the output does not match the prime in the command."
-    }
-
-    override fun verify(transaction: UtxoLedgerTransaction) {
-        val command = transaction.getCommands(PrimeCommands::class.java).singleOrNull() ?: throw CordaRuntimeException(REQUIRE_SINGLE_COMMAND)
-
-        when(command) {
-            is PrimeCommands.Query -> {
-                // This checks that the prime number generated by the service is greater than 0
-                val output = transaction.getOutputStates(Query::class.java).first()
-                NON_POSITIVE using { output.nthPrime > 0}
-            }
-            is PrimeCommands.Create -> {
-                val output = transaction.getOutputStates(Prime::class.java).first()
-                NON_MATCHING using {
-                    // This checks that the information in the command matches the state
-                    // This does NOT check that the Nth prime is correct
-                    (command.n == output.n && command.nthPrime == output.nthPrime)
-                }
-            }
-            else -> {
-                throw IllegalArgumentException("${UNRECOGNISED_COMMAND}: ${command::class.java.name}")
-            }
-        }
-    }
-
-    private infix fun String.using(expr: () -> Boolean) {
-        if (!expr.invoke()) throw CordaRuntimeException("Failed requirement: $this")
-    }
-}
